+++ DofusProtocol/Types/Types/game/character/alignment/ActorExtendedAlignmentInformations.cs
@@ -1,6 +1,6 @@
 
 
-// Generated on 11/16/2015 14:20:20
+// Generated on 12/20/2015 17:30:54
 using System;
 using System.Collections.Generic;
 using System.Linq;
@@ -26,7 +26,7 @@ namespace Stump.DofusProtocol.Types
         {
         }
         
-        public ActorExtendedAlignmentInformations(sbyte alignmentSide, sbyte alignmentValue, sbyte alignmentGrade, int characterPower, short honor, short honorGradeFloor, short honorNextGradeFloor, sbyte aggressable)
+        public ActorExtendedAlignmentInformations(sbyte alignmentSide, sbyte alignmentValue, sbyte alignmentGrade, double characterPower, short honor, short honorGradeFloor, short honorNextGradeFloor, sbyte aggressable)
          : base(alignmentSide, alignmentValue, alignmentGrade, characterPower)
         {
             this.honor = honor;
