+++ DofusProtocol/Types/Types/game/character/choice/AbstractCharacterToRefurbishInformation.cs
@@ -1,6 +1,6 @@
 
 
-// Generated on 11/16/2015 14:20:20
+// Generated on 12/20/2015 17:30:54
 using System;
 using System.Collections.Generic;
 using System.Linq;
@@ -24,7 +24,7 @@ namespace Stump.DofusProtocol.Types
         {
         }
         
-        public AbstractCharacterToRefurbishInformation(int id, IEnumerable<int> colors, int cosmeticId)
+        public AbstractCharacterToRefurbishInformation(long id, IEnumerable<int> colors, int cosmeticId)
          : base(id)
         {
             this.colors = colors;
