+++ DofusProtocol/Types/Types/game/context/fight/FightTeamMemberWithAllianceCharacterInformations.cs
@@ -1,6 +1,6 @@
 
 
-// Generated on 11/16/2015 14:20:21
+// Generated on 12/20/2015 17:30:56
 using System;
 using System.Collections.Generic;
 using System.Linq;
@@ -23,7 +23,7 @@ namespace Stump.DofusProtocol.Types
         {
         }
         
-        public FightTeamMemberWithAllianceCharacterInformations(int id, string name, byte level, Types.BasicAllianceInformations allianceInfos)
+        public FightTeamMemberWithAllianceCharacterInformations(double id, string name, byte level, Types.BasicAllianceInformations allianceInfos)
          : base(id, name, level)
         {
             this.allianceInfos = allianceInfos;
